+++ probe-rs/src/architecture/arm/core/armv7m.rs
@@ -14,7 +14,7 @@ use crate::{
     },
     core::{
         Architecture, CoreInformation, CoreInterface, CoreRegisters, CoreStatus, HaltReason,
-        MemoryMappedRegister, RegisterId, RegisterValue,
+        MemoryMappedRegister, RegisterId, RegisterValue, VectorCatchCondition,
     },
     error::Error,
     memory::valid_32bit_address,
@@ -807,7 +807,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
             self.write_core_reg(XPSR.id(), (xpsr_value | XPSR_THUMB).into())?;
         }
 
-        self.reset_catch_set()?;
+        self.reset_catch_clear()?;
 
         // try to read the program counter
         let pc_value = self.read_core_reg(self.program_counter().into())?;
@@ -891,8 +891,7 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
 
     fn write_core_reg(&mut self, address: RegisterId, value: RegisterValue) -> Result<(), Error> {
         if self.state.current_state.is_halted() {
-            super::cortex_m::write_core_reg(&mut *self.memory, address, value.try_into()?)?;
-            Ok(())
+            super::cortex_m::write_core_reg(&mut *self.memory, address, value.try_into()?)
         } else {
             Err(Error::Arm(ArmError::CoreNotHalted))
         }
@@ -1072,7 +1071,44 @@ impl<'probe> CoreInterface for Armv7m<'probe> {
     fn debug_core_stop(&mut self) -> Result<(), Error> {
         self.sequence
             .debug_core_stop(&mut *self.memory, CoreType::Armv7m)?;
+        Ok(())
+    }
 
+    #[tracing::instrument(skip(self))]
+    fn enable_vector_catch(&mut self, condition: VectorCatchCondition) -> Result<(), Error> {
+        let mut dhcsr = Dhcsr(self.memory.read_word_32(Dhcsr::get_mmio_address())?);
+        dhcsr.set_c_debugen(true);
+        self.memory
+            .write_word_32(Dhcsr::get_mmio_address(), dhcsr.into())?;
+
+        let mut demcr = Demcr(self.memory.read_word_32(Demcr::get_mmio_address())?);
+        match condition {
+            VectorCatchCondition::HardFault => demcr.set_vc_harderr(true),
+            VectorCatchCondition::CoreReset => demcr.set_vc_corereset(true),
+            VectorCatchCondition::All => {
+                demcr.set_vc_harderr(true);
+                demcr.set_vc_corereset(true);
+            }
+        };
+
+        self.memory
+            .write_word_32(Demcr::get_mmio_address(), demcr.into())?;
+        Ok(())
+    }
+
+    fn disable_vector_catch(&mut self, condition: VectorCatchCondition) -> Result<(), Error> {
+        let mut demcr = Demcr(self.memory.read_word_32(Demcr::get_mmio_address())?);
+        match condition {
+            VectorCatchCondition::HardFault => demcr.set_vc_harderr(false),
+            VectorCatchCondition::CoreReset => demcr.set_vc_corereset(false),
+            VectorCatchCondition::All => {
+                demcr.set_vc_harderr(false);
+                demcr.set_vc_corereset(false);
+            }
+        };
+
+        self.memory
+            .write_word_32(Demcr::get_mmio_address(), demcr.into())?;
         Ok(())
     }
 }