+++ probe-rs/src/core/registers.rs
@@ -22,18 +22,29 @@ pub enum RegisterDataType {
 /// while the [`CoreRegister::name`] will contain the architecture specific label of the register.
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum RegisterRole {
+    /// The default role for a register, with the name as defined by the architecture.
+    Core(&'static str),
     /// Function Argument registers like "A0", "a1", etc. (uses architecture specific names)
     Argument(&'static str),
     /// Function Return value registers like "R0", "r1", etc. (uses architecture specific names)
     Return(&'static str),
+    /// Program Counter register
     ProgramCounter,
+    /// Frame Pointer register
     FramePointer,
+    /// Stack Pointer register
     StackPointer,
+    /// Main Stack Pointer register
     MainStackPointer,
+    /// Process Stack Pointer register
     ProcessStackPointer,
+    /// Processor Status register
     ProcessorStatus,
+    /// Return Address register
     ReturnAddress,
+    /// Floating Point Unit register
     FloatingPoint,
+    /// Floating Point Status register
     FloatingPointStatus,
     /// Other architecture specific roles, e.g. "saved", "temporary", "variable", etc.
     Other(&'static str),
@@ -42,6 +53,7 @@ pub enum RegisterRole {
 impl Display for RegisterRole {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
+            RegisterRole::Core(name) => write!(f, "{}", name),
             RegisterRole::Argument(name) => write!(f, "{}", name),
             RegisterRole::Return(name) => write!(f, "{}", name),
             RegisterRole::ProgramCounter => write!(f, "PC"),
@@ -58,22 +70,52 @@ impl Display for RegisterRole {
     }
 }
 
+/// The rule used to preserve the value of a register between function calls duing unwinding,
+/// when DWARF unwind information is not available.
+///
+/// The rules for these are based on the 'Procedure Calling Standard' for each of the supported architectures:
+/// - Implemented: [AAPCS32](https://github.com/ARM-software/abi-aa/blob/main/aapcs32/aapcs32.rst#core-registers)
+/// - To be Implemented: [AAPCS64](https://github.com/ARM-software/abi-aa/blob/main/aapcs32/aapcs32.rst#core-registers)
+/// - To be Implemented: [RISC-V PCS](https://github.com/riscv-non-isa/riscv-elf-psabi-doc/releases/download/v1.0/riscv-abi.pdf)
+///
+/// Please note that the `Procedure Calling Standard` define register rules for the act of calling and/or returning from functions,
+/// while the timing of a stack unwinding is different (the `callee` has not yet completed / executed the epilogue),
+/// and the rules about preserving register values have to take this into account.
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub enum UnwindRule {
+    /// Callee-saved, a.k.a non-volatile registers, or call-preserved.
+    /// If there is DWARF unwind `RegisterRule` we will apply it during unwind,
+    /// otherwise we assume it was untouched and preserve the current value.
+    Preserve,
+    /// Caller-saved, a.k.a. volatile registers, or call-clobbered.
+    /// If there is DWARF unwind `RegisterRule` we will apply it during unwind,
+    /// otherwise we assume it was corrupted by the callee, and clear the value.
+    /// Note: This is the default value, and is used for all situations where DWARF unwind
+    /// information is not available, and the register is not explicitly marked in the definition.
+    #[default]
+    Clear,
+    /// Additional rules are required to determine the value of the register.
+    /// These are typically found in either the DWARF unwind information,
+    /// or requires additional platform specific registers to be read.
+    SpecialRule,
+}
+
 /// Describes a core (or CPU / hardware) register with its properties.
 /// Each architecture will have a set of general purpose registers, and potentially some special purpose registers. It also happens that some general purpose registers can be used for special purposes. For instance, some ARM variants allows the `LR` (link register / return address) to be used as general purpose register `R14`."
 #[derive(Debug, Clone, PartialEq)]
 pub struct CoreRegister {
-    /// The architecture specific name of the register. This may be identical, or similar to [`RegisterRole`].
-    pub(crate) name: &'static str,
     // /// Some architectures have multiple names for the same register, depending on the context and the role of the register.
     pub(crate) id: RegisterId,
     /// If the register plays a special role (one or more) during program execution and exception handling, this array will contain the appropriate [`RegisterRole`] entry/entries.
     pub(crate) roles: &'static [RegisterRole],
     pub(crate) data_type: RegisterDataType,
+    /// For unwind purposes (debug and/or exception handling), we need to know how values are preserved between function calls. (Applies to ARM and RISC-V)
+    pub unwind_rule: UnwindRule,
 }
 
 impl Display for CoreRegister {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let primary_name = self.name;
+        let primary_name = self.name();
         write!(f, "{}", primary_name)?;
         if !self.roles.is_empty() {
             for role in self.roles {
@@ -87,9 +129,15 @@ impl Display for CoreRegister {
 }
 
 impl CoreRegister {
-    /// Get the display name of this register
+    /// Get the primary display name (As defined by `RegisterRole::Core()` of this register
     pub fn name(&self) -> &'static str {
-        self.name
+        self.roles
+            .iter()
+            .find_map(|role| match role {
+                RegisterRole::Core(name) => Some(*name),
+                _ => None,
+            })
+            .unwrap_or("Unknown")
     }
 
     /// Get the id of this register
@@ -378,7 +426,7 @@ impl CoreRegisters {
         CoreRegisters(core_registers)
     }
 
-    /// Returns an iterator over the descriptions of all the core registers (non-FPU) of this core.
+    /// Returns an iterator over the descriptions of all the non-FPU registers of this core.
     pub fn core_registers(&self) -> impl Iterator<Item = &CoreRegister> {
         self.0
             .iter()
@@ -391,6 +439,11 @@ impl CoreRegisters {
                 })
             })
             .cloned()
+    }
+
+    /// Returns an iterator over the descriptions of all the registers of this core.
+    pub fn all_registers(&self) -> impl Iterator<Item = &CoreRegister> {
+        self.0.iter().cloned()
     }
 
     /// Returns the nth platform register.
@@ -449,6 +502,14 @@ impl CoreRegisters {
             })
             .cloned()
             .nth(index)
+    }
+
+    /// The program counter.
+    pub fn pc(&self) -> Option<&CoreRegister> {
+        self.0
+            .iter()
+            .find(|r| r.register_has_role(RegisterRole::ProgramCounter))
+            .cloned()
     }
 
     /// The main stack pointer.