+++ probe-rs/src/probe/espusbjtag/protocol.rs
@@ -3,16 +3,17 @@ use std::{fmt::Debug, time::Duration};
 use rusb::{request_type, Context, Device, Direction, TransferType, UsbContext};
 
 use crate::{
-    probe::common::BitIter, DebugProbeError, DebugProbeInfo, DebugProbeSelector, DebugProbeType,
-    ProbeCreationError,
+    probe::common::OwnedBitIter, DebugProbeError, DebugProbeInfo, DebugProbeSelector,
+    DebugProbeType, ProbeCreationError,
 };
 
 const JTAG_PROTOCOL_CAPABILITIES_VERSION: u8 = 1;
 const JTAG_PROTOCOL_CAPABILITIES_SPEED_APB_TYPE: u8 = 1;
 const MAX_COMMAND_REPETITIONS: usize = 1024;
-const OUT_BUFFER_SIZE: usize = OUT_EP_BUFFER_SIZE * 64;
-const OUT_EP_BUFFER_SIZE: usize = 64;
+const OUT_BUFFER_SIZE: usize = OUT_EP_BUFFER_SIZE * 32;
+const OUT_EP_BUFFER_SIZE: usize = 128;
 const IN_EP_BUFFER_SIZE: usize = 64;
+const HW_FIFO_SIZE: usize = 4;
 const USB_TIMEOUT: Duration = Duration::from_millis(5000);
 const USB_DEVICE_CLASS: u8 = 0xFF;
 const USB_DEVICE_SUBCLASS: u8 = 0xFF;
@@ -37,12 +38,16 @@ pub(super) struct ProtocolHandler {
     // a mechanism to compress the datastream by adding a `Repeat` command to repeat the previous command `n` times instead of
     // actually putting the command into the queue `n` times.
     output_buffer: Vec<Command>,
-    // A store for all the read bits (from the traget) such that the BitIter the methods return can borrow and iterate over it.
-    input_buffer: Vec<u8>,
+    // A store for all the read bits (from the target) such that the BitIter the methods return can borrow and iterate over it.
+    input_buffers: Vec<OwnedBitIter>,
     pending_in_bits: usize,
 
     ep_out: u8,
     ep_in: u8,
+
+    pub(crate) base_speed_khz: u32,
+    pub(crate) div_min: u16,
+    pub(crate) div_max: u16,
 }
 
 impl Debug for ProtocolHandler {
@@ -50,9 +55,12 @@ impl Debug for ProtocolHandler {
         f.debug_struct("ProtocolHandler")
             .field("command_queue", &self.command_queue)
             .field("output_buffer", &self.output_buffer)
-            .field("input_buffer", &self.input_buffer)
+            .field("input_buffers", &self.input_buffers)
             .field("ep_out", &self.ep_out)
             .field("ep_in", &self.ep_in)
+            .field("base_speed_khz", &self.base_speed_khz)
+            .field("div_min", &self.div_min)
+            .field("div_max", &self.div_max)
             .finish()
     }
 }
@@ -163,10 +171,9 @@ impl ProtocolHandler {
             USB_TIMEOUT,
         )?;
 
-        // TODO:
-        // let mut base_speed_khz = 1000;
-        // let mut div_min = 1;
-        // let mut div_max = 1;
+        let mut base_speed_khz = 1000;
+        let mut div_min = 1;
+        let mut div_max = 1;
 
         let protocol_version = buffer[0];
         tracing::debug!("{:?}", &buffer[..20]);
@@ -185,10 +192,11 @@ impl ProtocolHandler {
             let length = buffer[p + 1];
 
             if typ == JTAG_PROTOCOL_CAPABILITIES_SPEED_APB_TYPE {
-                // TODO:
-                // base_speed_khz = (((buffer[p + 3] as u16) << 8) | buffer[p + 2] as u16) * 10 / 2;
-                // div_min = ((buffer[p + 5] as u16) << 8) | buffer[p + 4] as u16;
-                // div_max = ((buffer[p + 7] as u16) << 8) | buffer[p + 6] as u16;
+                base_speed_khz =
+                    ((((buffer[p + 3] as u16) << 8) | buffer[p + 2] as u16) as u64 * 10 / 2) as u32;
+                div_min = ((buffer[p + 5] as u16) << 8) | buffer[p + 4] as u16;
+                div_max = ((buffer[p + 7] as u16) << 8) | buffer[p + 6] as u16;
+                tracing::info!("Found ESP USB JTAG adapter, base speed is {base_speed_khz}khz. Available dividers: ({div_min}..{div_max})");
             } else {
                 tracing::warn!("Unknown capabilities type {:01X?}", typ);
             }
@@ -196,20 +204,21 @@ impl ProtocolHandler {
             p += length as usize;
         }
 
-        // TODO:
-        // let hw_in_fifo_len = 4;
-
         tracing::debug!("Succesfully attached to ESP USB JTAG.");
 
         Ok(Self {
             device_handle,
             command_queue: None,
             output_buffer: Vec::new(),
-            input_buffer: Vec::new(),
+            input_buffers: Vec::new(),
             // The following expects are okay as we check that the values we call them on are `Some`.
             ep_out: ep_out.expect("This is a bug. Please report it."),
             ep_in: ep_in.expect("This is a bug. Please report it."),
             pending_in_bits: 0,
+
+            base_speed_khz,
+            div_min,
+            div_max,
         })
     }
 
@@ -219,7 +228,21 @@ impl ProtocolHandler {
         tms: impl IntoIterator<Item = bool>,
         tdi: impl IntoIterator<Item = bool>,
         cap: bool,
-    ) -> Result<BitIter, DebugProbeError> {
+    ) -> Result<OwnedBitIter, DebugProbeError> {
+        self.jtag_io_async(tms, tdi, cap)?;
+        self.flush()
+    }
+
+    /// Put a bit on TDI and possibly read one from TDO.
+    /// to recieve the bytes from this operations call [`ProtocolHandler::flush`]
+    ///
+    /// Note that if the internal buffer is exceeded bytes will be automatically flushed to usb device
+    pub fn jtag_io_async(
+        &mut self,
+        tms: impl IntoIterator<Item = bool>,
+        tdi: impl IntoIterator<Item = bool>,
+        cap: bool,
+    ) -> Result<(), DebugProbeError> {
         tracing::debug!("JTAG IO! {} ", cap);
         for (tms, tdi) in tms.into_iter().zip(tdi.into_iter()) {
             self.push_command(Command::Clock { cap, tdi, tms })?;
@@ -227,8 +250,7 @@ impl ProtocolHandler {
                 self.pending_in_bits += 1;
             }
         }
-
-        self.flush()
+        Ok(())
     }
 
     /// Sets the two different resets on the target.
@@ -260,14 +282,13 @@ impl ProtocolHandler {
     }
 
     /// Flushes all the pending commands to the JTAG adapter.
-    pub fn flush(&mut self) -> Result<BitIter, DebugProbeError> {
+    pub fn flush(&mut self) -> Result<OwnedBitIter, DebugProbeError> {
         if let Some((command_in_queue, repetitions)) = self.command_queue.take() {
             self.write_stream(command_in_queue, repetitions)?;
         }
 
         tracing::debug!("Flushing ...");
 
-        // https://github.com/espressif/openocd-esp32/blob/a28f71785066722f49494e0d946fdc56966dcc0d/src/jtag/drivers/esp_usb_jtag.c#L423
         self.add_raw_command(Command::Flush)?;
 
         // Make sure we add an additional nibble to the command buffer if the number of nibbles is odd,
@@ -278,7 +299,19 @@ impl ProtocolHandler {
 
         self.send_buffer()?;
 
-        self.receive_buffer()
+        while self.pending_in_bits != 0 {
+            self.receive_buffer()?;
+        }
+
+        let iter = self
+            .input_buffers
+            .iter()
+            .flat_map(|it| it.clone())
+            .collect();
+
+        self.input_buffers.clear();
+
+        Ok(iter)
     }
 
     /// Writes a command one or multiple times into the raw buffer we send to the USB EP later
@@ -323,16 +356,20 @@ impl ProtocolHandler {
             self.send_buffer()?;
         }
 
-        // TODO: Bizarre undocumented condition to flush buffer.
-        // https://github.com/espressif/openocd-esp32/blob/a28f71785066722f49494e0d946fdc56966dcc0d/src/jtag/drivers/esp_usb_jtag.c#L367
+        // Undocumented condition to flush buffer.
+        // First check, whether the output buffer is suitable for flushing? it should be modulo of the EP buffer size
+        if self.output_buffer.len() % OUT_EP_BUFFER_SIZE == 0 {
+            // Second check, if it is suitable, is there enough to flush
+            if self.pending_in_bits > (IN_EP_BUFFER_SIZE + HW_FIFO_SIZE) * 8 {
+                self.send_buffer()?;
+            }
+        }
 
         Ok(())
     }
 
     /// Sends the commands stored in the output buffer to the USB EP.
     fn send_buffer(&mut self) -> Result<(), DebugProbeError> {
-        tracing::trace!("Command Buffer: {:?}", self.output_buffer);
-
         let commands = self
             .output_buffer
             .chunks(2)
@@ -347,80 +384,93 @@ impl ProtocolHandler {
             })
             .collect::<Vec<_>>();
 
-        self.device_handle
-            .write_bulk(self.ep_out, &commands, USB_TIMEOUT)
-            .map_err(|e| DebugProbeError::Usb(Some(Box::new(e))))?;
+        tracing::trace!(
+            "Writing {}byte ({}nibles) to usb endpoint",
+            commands.len(),
+            commands.len() * 2
+        );
+        let mut offset = 0;
+        let mut total = 0;
+        loop {
+            let bytes = self
+                .device_handle
+                .write_bulk(self.ep_out, &commands[offset..], USB_TIMEOUT)
+                .map_err(|e| DebugProbeError::Usb(Some(Box::new(e))))?;
+            total += bytes;
+            offset += bytes;
+
+            if total == commands.len() {
+                break;
+            }
+        }
 
         // We only clear the output buffer on a successful transmission of all bytes.
         self.output_buffer.clear();
 
-        // TODO: Weird condition again.
-        // https://github.com/espressif/openocd-esp32/blob/a28f71785066722f49494e0d946fdc56966dcc0d/src/jtag/drivers/esp_usb_jtag.c#L345
+        // If there's more than a bufferful of data queing up in the jtag adapters IN endpoint, empty all but one buffer.
+        loop {
+            if self.pending_in_bits > (IN_EP_BUFFER_SIZE + HW_FIFO_SIZE) * 8 {
+                self.receive_buffer()?;
+            } else {
+                break;
+            }
+        }
 
         Ok(())
     }
 
     /// Tries to receive pending in bits from the USB EP.
-    fn receive_buffer(&mut self) -> Result<BitIter, DebugProbeError> {
-        self.input_buffer = vec![0; (self.pending_in_bits + 7) / 8];
+    fn receive_buffer(&mut self) -> Result<(), DebugProbeError> {
+        let count = ((self.pending_in_bits + 7) / 8).min(IN_EP_BUFFER_SIZE);
+        let mut incoming = vec![0; count];
 
-        let mut bits_read = 0;
+        tracing::trace!("Receiving buffer, pending bits: {}", self.pending_in_bits);
 
-        while bits_read != self.pending_in_bits {
-            let count = ((self.pending_in_bits + 7) / 8).min(IN_EP_BUFFER_SIZE);
-            tracing::trace!("Receiveing {} bytes.", count);
-
-            if count == 0 {
-                return Ok(BitIter::new(&[], 0));
-            }
-
-            let offset = bits_read / 8;
-
-            let mut tries = 3;
-            let mut read_bytes;
-            loop {
-                read_bytes = self
-                    .device_handle
-                    .read_bulk(
-                        self.ep_in,
-                        &mut self.input_buffer[offset..offset + count],
-                        USB_TIMEOUT,
-                    )
-                    .map_err(|e| {
-                        tracing::warn!("Something went wrong in read_bulk {:?}", e);
-                        DebugProbeError::Usb(Some(Box::new(e)))
-                    })?;
-
-                tracing::trace!("Read bytes: {} bytes. On try {}", read_bytes, 4 - tries);
-
-                if read_bytes != 0 {
-                    break;
-                }
-
-                tries -= 1;
-
-                if tries == 0 {
-                    break;
-                }
-            }
-
-            if read_bytes == 0 {
-                // Sometimes the hardware returns 0 bytes instead of NAKing the transaction. Ignore this.
-                return Err(DebugProbeError::ProbeSpecific(
-                    "Transaction not acknowledged.".into(),
-                ));
-            }
-
-            tracing::trace!("Received {} bytes.", count);
-
-            let bits_in_buffer = self.pending_in_bits.min(count * 8);
-            bits_read += bits_in_buffer;
+        if count == 0 {
+            return Ok(());
         }
 
-        tracing::trace!("Read: {:?}, length = {}", self.input_buffer, bits_read);
-        self.pending_in_bits -= bits_read;
+        let mut offset = 0;
+        let mut total = 0;
+        loop {
+            let read_bytes = self
+                .device_handle
+                .read_bulk(self.ep_in, &mut incoming[offset..], USB_TIMEOUT)
+                .map_err(|e| {
+                    tracing::warn!(
+                        "Something went wrong in read_bulk {:?} when trying to read {}bytes - pending_in_bits: {}",
+                        e,
+                        count,
+                        self.pending_in_bits,
+                    );
+                    DebugProbeError::Usb(Some(Box::new(e)))
+                })?;
+            total += read_bytes;
+            offset += read_bytes;
 
-        Ok(BitIter::new(&self.input_buffer, bits_read))
+            if read_bytes == 0 {
+                tracing::debug!("Read 0 bytes from USB");
+                return Ok(());
+            }
+
+            if total == count {
+                break;
+            } else {
+                tracing::warn!("USB only recieved {} out of {} bytes", read_bytes, count);
+            }
+
+            tracing::trace!("Received {} bytes.", read_bytes);
+        }
+
+        let bits_in_buffer = self.pending_in_bits.min(total * 8);
+
+        tracing::trace!("Read: {:?}, length = {}", incoming, bits_in_buffer);
+        self.pending_in_bits -= bits_in_buffer;
+
+        self.input_buffers
+            .push(OwnedBitIter::new(&incoming, bits_in_buffer));
+
+        Ok(())
     }
 }